Caching is more trouble than it is worth
Setting the `Cache-Control` header to `ONE_DAY` meant that it was being reset
every time a user visited the site. The goal was really to have the
`Cache-Control` set to "the end of today", rather than "one day from this
second".

That's easily implemented, but is honestly more trouble than it's worth, so I'm
just removing caching.